% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MetSSS_model.R
\docType{data}
\name{MetSSS_model}
\alias{MetSSS_model}
\title{MetSSS Model}
\format{an object from \code{\link{mahalanobisComposite}}}
\source{
Healthy Hearts Study
}
\usage{
MetSSS_model
}
\description{
Simple example model that predicts hours of watching TV from age and marital status
}

